requirements.txt
alembic==0.9.6
Babel==2.5.1
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dominate==2.3.1
elasticsearch==6.1.1
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Excel==0.0.7
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
lml==0.0.4
Mako==1.0.7
MarkupSafe==1.0
openpyxl==2.5.9
pdfkit==0.6.1
pyecharts-jupyter-installer==0.0.3
pyexcel==0.5.9.1
pyexcel-handsontable==0.0.2
pyexcel-io==0.5.9.1
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
PyJWT==1.5.3
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
redis==2.10.6
requests==2.18.4
rq==0.9.2
six==1.11.0
SQLAlchemy==1.1.14
texttable==1.4.0
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
xlrd==1.1.0
xlwt==1.3.0
